+++ src/mc_test_circuits/csw.rs
@@ -167,9 +167,9 @@ pub fn generate_parameters(
     ps: ProvingSystem,
     num_constraints: u32,
     with_constant: bool,
-    segment_size: Option<usize>,
+    segment_size: Option<u32>,
 ) -> Result<(ZendooProverKey, ZendooVerifierKey), ProvingSystemError> {
-    let supported_degree = segment_size.map(|ss| ss - 1 );
+    let supported_degree = segment_size.map(|ss| (ss - 1) as usize);
     let ck_g1 = get_g1_committer_key(supported_degree)?;
     match ps {
         ProvingSystem::Undefined => Err(ProvingSystemError::UndefinedProvingSystem),
@@ -220,10 +220,10 @@ pub fn generate_proof(
     cert_data_hash: &FieldElement,
     end_cumulative_sc_tx_commitment_tree_root: &FieldElement,
     num_constraints: u32,
-    segment_size: Option<usize>,
+    segment_size: Option<u32>,
 ) -> Result<ZendooProof, ProvingSystemError> {
     let rng = &mut OsRng;
-    let supported_degree = segment_size.map(|ss| ss - 1 );
+    let supported_degree = segment_size.map(|ss| (ss - 1) as usize);
     let ck_g1 = get_g1_committer_key(supported_degree)?;
 
     let mut fes = DataAccumulator::init()
